requirements.txt
APScheduler==3.11.0
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2024.8.30
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.0
dateparser==1.2.0
Django==5.1.1
django-apscheduler==0.7.0
django-cors-headers==4.4.0
django-dbbackup==4.3.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.10
imageio==2.36.0
incremental==24.7.2
lazy_loader==0.4
moment==0.12.1
msgpack==1.1.1
networkx==3.4.2
numpy==2.1.1
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==10.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
razorpay==1.4.2
redis==5.2.1
regex==2024.9.11
requests==2.32.3
rest-framework-simplejwt==0.0.2
scikit-image==0.24.0
scipy==1.14.1
service-identity==24.2.0
setuptools==80.9.0
six==1.16.0
sqlparse==0.5.1
tifffile==2024.9.20
times==0.7
Twisted==25.5.0
txaio==25.6.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.14.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.3
XlsxWriter==3.2.1
zope.interface==7.2
